gpu-backend/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
numpy>=1.26.0
controlnet-aux>=0.0.7
peft>=0.9.0
